Fixed Issues and Removed Unnecessary File

+++ app/core/functions.php
@@ -110,8 +110,9 @@ function generateProfilePicture($initials) {
 	$fontSize = 40;
 	$bbox = imagettfbbox($fontSize, 0, 'assets/fonts/Europa-Regular.ttf', $initials);
 	$textWidth = $bbox[2] - $bbox[0];
-	$x = ($width - $textWidth) / 2;
-	$y = ($height + $fontSize) / 2;
+	// Cast to integer if needed
+	$x = (int)(($width - $textWidth) / 2);
+	$y = (int)(($height + $fontSize) / 2);
 	
 	// Add the initials to the image
 	imagettftext($image, $fontSize, 0, $x, $y, $textColor, 'assets/fonts/Europa-Regular.ttf', $initials);
